Makefile: don't require `which`
`which` is provided by "debianutils" package or "which" package,
which may not be installed by default.

It's easy to add "which" or "debianutils" as a build-time dependency, but
generally, there should be no need to resolve tool names into full paths beforehand --
shell should be able to do it on its own.
(given that the PATH is setup correctly)

For `git` and `pandoc`, `which` was used to determine if they were installed/available.
Instead, try to call them and see if they report their versions.
(or we could ask shell to resolve paths to them,
 but this may not be portable across different shells)

Originally suggested by @mgorny [1].

[1] https://github.com/gentoo/gentoo/pull/9544#issuecomment-414146058